# apps/web/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "CostLedgerEntry": {
        "properties": {
          "amount_usd": {
            "title": "Amount Usd",
            "type": "string"
          },
          "cost_type": {
            "title": "Cost Type",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "pricing_version": {
            "title": "Pricing Version",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "cost_type",
          "amount_usd",
          "pricing_version"
        ],
        "title": "CostLedgerEntry",
        "type": "object"
      },
      "CostSummaryResponse": {
        "properties": {
          "concurrency_note": {
            "title": "Concurrency Note",
            "type": "string"
          },
          "incident_ref": {
            "title": "Incident Ref",
            "type": "string"
          },
          "ledger": {
            "items": {
              "$ref": "#/components/schemas/CostLedgerEntry"
            },
            "title": "Ledger",
            "type": "array"
          },
          "model_calls": {
            "title": "Model Calls",
            "type": "integer"
          },
          "model_cost": {
            "title": "Model Cost",
            "type": "string"
          },
          "pricing_versions": {
            "items": {
              "type": "string"
            },
            "title": "Pricing Versions",
            "type": "array"
          },
          "tool_calls": {
            "title": "Tool Calls",
            "type": "integer"
          },
          "tool_cost": {
            "title": "Tool Cost",
            "type": "string"
          },
          "total_cost": {
            "title": "Total Cost",
            "type": "string"
          }
        },
        "required": [
          "incident_ref",
          "model_cost",
          "tool_cost",
          "total_cost",
          "model_calls",
          "tool_calls",
          "pricing_versions",
          "concurrency_note",
          "ledger"
        ],
        "title": "CostSummaryResponse",
        "type": "object"
      },
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ref"
          },
          "resource": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resource"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "EvaluationResponse": {
        "properties": {
          "evaluation_cost": {
            "title": "Evaluation Cost",
            "type": "string"
          },
          "evaluator_version": {
            "title": "Evaluator Version",
            "type": "string"
          },
          "llm_judge_used": {
            "title": "Llm Judge Used",
            "type": "boolean"
          },
          "passed": {
            "title": "Passed",
            "type": "integer"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/EvaluationResultItem"
            },
            "title": "Results",
            "type": "array"
          },
          "suite_name": {
            "title": "Suite Name",
            "type": "string"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "suite_name",
          "evaluator_version",
          "passed",
          "total",
          "llm_judge_used",
          "evaluation_cost",
          "results"
        ],
        "title": "EvaluationResponse",
        "type": "object"
      },
      "EvaluationResultItem": {
        "properties": {
          "actual": {
            "title": "Actual",
            "type": "string"
          },
          "check_name": {
            "title": "Check Name",
            "type": "string"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "expected": {
            "title": "Expected",
            "type": "string"
          },
          "outcome": {
            "title": "Outcome",
            "type": "string"
          }
        },
        "required": [
          "check_name",
          "outcome",
          "expected",
          "actual",
          "detail"
        ],
        "title": "EvaluationResultItem",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InterventionView": {
        "properties": {
          "action_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Action Status"
          },
          "action_type": {
            "title": "Action Type",
            "type": "string"
          },
          "composite_score": {
            "title": "Composite Score",
            "type": "string"
          },
          "decision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision"
          },
          "executed": {
            "title": "Executed",
            "type": "boolean"
          },
          "expected_value": {
            "title": "Expected Value",
            "type": "string"
          },
          "integration_status": {
            "title": "Integration Status",
            "type": "string"
          },
          "matched_rules": {
            "items": {
              "type": "string"
            },
            "title": "Matched Rules",
            "type": "array"
          },
          "rank": {
            "title": "Rank",
            "type": "integer"
          },
          "rationale": {
            "title": "Rationale",
            "type": "string"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "risk_tier": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Tier"
          },
          "target_ref": {
            "title": "Target Ref",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "rank",
          "title",
          "action_type",
          "rationale",
          "target_ref",
          "expected_value",
          "composite_score",
          "decision",
          "risk_tier",
          "matched_rules",
          "reason",
          "executed",
          "action_status",
          "integration_status"
        ],
        "title": "InterventionView",
        "type": "object"
      },
      "InvestigationResponse": {
        "properties": {
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidenceItemView"
            },
            "title": "Evidence",
            "type": "array"
          },
          "hypotheses": {
            "items": {
              "$ref": "#/components/schemas/HypothesisView"
            },
            "title": "Hypotheses",
            "type": "array"
          },
          "impact": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ImpactView"
              },
              {
                "type": "null"
              }
            ]
          },
          "incident_ref": {
            "title": "Incident Ref",
            "type": "string"
          }
        },
        "required": [
          "incident_ref",
          "evidence",
          "hypotheses",
          "impact"
        ],
        "title": "InvestigationResponse",
        "type": "object"
      },
...
      "TimelineResponse": {
        "properties": {
          "events": {
            "items": {
              "$ref": "#/components/schemas/TimelineEventView"
            },
            "title": "Events",
            "type": "array"
          },
          "incident_ref": {
            "title": "Incident Ref",
            "type": "string"
          },
          "trace_count": {
            "title": "Trace Count",
            "type": "integer"
          }
        },
        "required": [
          "incident_ref",
          "trace_count",
          "events"
        ],
        "title": "TimelineResponse",
        "type": "object"
      },
...
      "UncertainActionsResponse": {
        "properties": {
          "actions": {
            "items": {
              "$ref": "#/components/schemas/UncertainActionView"
            },
            "title": "Actions",
            "type": "array"
          },
          "delivery_note": {
            "title": "Delivery Note",
            "type": "string"
          },
          "incident_ref": {
            "title": "Incident Ref",
            "type": "string"
          }
        },
        "required": [
          "incident_ref",
          "actions",
          "delivery_note"
        ],
        "title": "UncertainActionsResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Agentic AI GTM Control Tower. Every external integration is SIMULATED in v1 -- see CAPABILITY_MATRIX.md.",
    "title": "Revenue Sentinel",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/evaluation/latest": {
      "get": {
        "description": "The most recent evaluation attempt. History is append-only, so this is a view of\none attempt rather than a mutable status.",
        "operationId": "evaluation_evaluation_latest_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          }
        },
        "summary": "Evaluation",
        "tags": [
          "dashboard"
        ]
      }
    },
...
    "/incidents/{incident_ref}/cost": {
      "get": {
        "description": "Microdollar precision throughout -- `$0.000000` is not rounded to `$0.00`.",
        "operationId": "cost_incidents__incident_ref__cost_get",
        "parameters": [
          {
            "in": "path",
            "name": "incident_ref",
            "required": true,
            "schema": {
              "title": "Incident Ref",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CostSummaryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cost",
        "tags": [
          "dashboard"
        ]
      }
    },
    "/incidents/{incident_ref}/interventions": {
      "get": {
        "description": "Ranked interventions with the decision and the rules behind it.",
        "operationId": "interventions_incidents__incident_ref__interventions_get",
        "parameters": [
          {
            "in": "path",
            "name": "incident_ref",
            "required": true,
            "schema": {
              "title": "Incident Ref",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/InterventionView"
                  },
                  "title": "Response Interventions Incidents  Incident Ref  Interventions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Interventions",
        "tags": [
          "dashboard"
        ]
      }
    },
    "/incidents/{incident_ref}/investigation": {
      "get": {
        "description": "Evidence, hypotheses and the deterministic impact figures.",
        "operationId": "investigation_incidents__incident_ref__investigation_get",
        "parameters": [
          {
            "in": "path",
            "name": "incident_ref",
            "required": true,
            "schema": {
              "title": "Incident Ref",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestigationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Investigation",
        "tags": [
          "dashboard"
        ]
      }
    },
    "/incidents/{incident_ref}/timeline": {
      "get": {
        "description": "The trace-correlated merge. Absent tracing metadata stays absent.",
        "operationId": "timeline_incidents__incident_ref__timeline_get",
        "parameters": [
          {
            "in": "path",
            "name": "incident_ref",
            "required": true,
            "schema": {
              "title": "Incident Ref",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimelineResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Timeline",
        "tags": [
          "dashboard"
        ]
      }
    },
    "/incidents/{incident_ref}/uncertain-actions": {
      "get": {
        "description": "Actions whose outcome is unknown, and the command that resolves one.\n\nRead-only, like everything else here. Reconciliation is an accountable act by a named\nperson and there is no authenticated identity, so this renders the CLI command rather\nthan offering a button (ADR-0022, ADR-0025).",
        "operationId": "uncertain_actions_incidents__incident_ref__uncertain_actions_get",
        "parameters": [
          {
            "in": "path",
            "name": "incident_ref",
            "required": true,
            "schema": {
              "title": "Incident Ref",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UncertainActionsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Uncertain Actions",
        "tags": [
          "dashboard"
        ]
      }
    },
...
  }
}